switch go using childprocess for running the child process
