+++ lib/launchy.rb
@@ -157,3 +157,4 @@ require 'launchy/descendant_tracker'
 require 'launchy/error'
 require 'launchy/application'
 require 'launchy/detect'
+require 'launchy/runner'
