+++ lib/launchy/detect.rb
@@ -6,4 +6,3 @@ end
 require 'launchy/detect/host_os'
 require 'launchy/detect/host_os_family'
 require 'launchy/detect/nix_desktop_environment'
-require 'launchy/detect/runner'
